agentcore-agents/AdoMantenimiento/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "AdoMantenimiento"
version = "0.1.0"
description = "Agente de Mantenimiento Predictivo - ADO MobilityIA"
requires-python = ">=3.10"
dependencies = [
    "aws-opentelemetry-distro",
    "bedrock-agentcore >= 1.0.3",
    "boto3 >= 1.35.0",
    "botocore[crt] >= 1.35.0",
    "strands-agents >= 1.13.0",
    "strands-agents-builder >= 0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
